backend/analysis/requirements.txt
psycopg2-binary==2.9.9
pypdf==4.0.1
python-docx==1.1.0